Use struct{}{} instead of defined var
The justification for this change is that 'struct{}{}' is immediately
obvious; it is exactly what it says it it. Conversely, the value of
'flag' needs to be found in its declaration; this is trivial if the
reader is aware of Go types and given knowledge of the map type obvious,
but naming the value adds nothing except possibly some subjective
prettiness, and is arguably incorrect since the value in 'flag' is not
actually the value that is stored and used as a flag.