Remove set package
Required methods are retained and unexported (this allows simplifying
the language to be more idiomatic Go since we don't need to coform to
set language), other methods are dropped, or kept in testing where
needed there.

The issue of using pointers to maps is skirted by having two types of
simple sets, simpleSet and set. simpleSet does not have a pointer
receiver since it never needs to make the kinds of comparisons the set
type does - Jeff's approach of using pointer equality for fast path
tests is an elegant approach is I don't think there is another simpler
or faster approach. Since we are not exporting the set type now, there
should be limited scope for abuse.

In PostDominators and Dominators the node lists are returned now as a
map[int][]int. This removes the requirement to export the set type and
also should improve performance for reuse of these lists by clients (I
should benchmark the impact). The lists are not sorted but can be by
clients.