Remove blas support for column major matrices
It was decided, after discussion on the gonum-dev list, to remove
support for column-major matrices in the blas package. This commit
completely removes all public support for the blas interface. Some
necessary residuals remain in the cblas package because the c functions
require specifying an order. However, these residuals are private and
are thus invisible to package users.

Several of the testing scripts tested column major implementation only.
This commit comments out those tests for the time being until they can
be migrated to row major.