Fix cholesky size comparison error, and add test. Remove redundant CholeskyR function
The current implementation of cholesky checks that the number of columns matches, not the number of rows. This means it panics for matrices of the correct size, and gives and out of bounds errors for matrices of incorrect size. This corrects the check.

There is also an untested CholeskyR function which is redundant for current
use cases. Remove it (for now at least).