Put returns back into [Op]To functions
Functions which always result in the same dot slice should return
slices, as agreed during discussion in gonum-dev.

I also added tests to make sure that the returned slices were the same
as the mutated slices.  Even though the coverage was reporting 100%,
the tests were not covering 100% of the returned values.