run go fmt on goblas
Ran go fmt ./… on the repo and only this file was changed.  Longer term
it should be fixed with changes to the script which builds it, but that
script currently points to a referenceblas file which doesn’t exist,
and cblas had go fmt ran manually on it, so this seems like a
reasonable thing to do.