Update .travis.yml
This travis file will build and test floats using both release and tip,
and then ensure that the code is properly formatted.  If any tests
fail, it will currently just email me.

The email list should be expanded, but I didn’t want to “volunteer”
people to keep an eye on this package.