correct weighted test, and make fixture clearer
Changed the names of the test table to be clearer.  Also made the
answer a matrix and used Dense.Equals instead of performing the
comparison on r, c, and the raw data.

Also, the old results for the weighted case were wrong.  I’ve added the
same items into the stat_test.go file.