Changed gradient sig to move gradient first.
The convention is to put receivers first, and allocate if nil. Previously, Gradient had not followed this convention. This moves the []float64 receiver to be the first argument, allocates it if nil, and returns the value from the function. This brings it into convention with the other packages. This PR also adds a test for the nil case.