Split Max and Min functions in two
This PR changes the behavior of Max to just return the maximum value, and creates MaxIdx for getting the index of the maximum value. A similar change is made for Min. Often times it is convenient to use these expressions as part of larger expressions, and this change makes floats more enjoyable to use