Remove extra check for conclusion of Bisection.
The original code for Bisection included a secondary check to the strong
Wolfe conditions to see if the optmiization was finished. The idea was
to help mitigate floating point noise and allow for stronger convergence
to the gradient. Unfortunately, all this does is add complexity. The
parameters are ad hoc, and trade off floating point noise for actual
function modulation. For more complicated functions (especially
concurrent ones) the noise will be higher, while other functions may
have modulations that are very small. It is impossible to design a
tradeoff that is good for all functions. Instead, keep the code simple.
This also fixes issues with Bisection and the outer OptLoc disagreeing
on the optimum location. Closes #53