Test initial location for NaN or Inf gradient values.
Bad initial gradient values cause the line searches to fail as the Wolfe conditions are never met. Test the initial gradient to make sure it is not NaN or Inf. While here, fix the GulfRD test which has a gradient that evaluates to NaN at the optimum location. This allows the test at the optimum to be re-enabled