A+C: Add Renée French for numerical gopher
Renée French did the numerical gopher in square brackets as a
commission, so it is in that regard distinct from the Go gopher. This
A+C addition is to reflect the fact that she put work into the gonum
project aside from her CC artwork.

The email address field of the A+C lines is empty since she has no
publically visible textual representation of her email address, and I
negelected to ask if it was OK to publish her email when asking if she
was OK having her name in the A+C.