Simplify evaluate() and invalidate()
Until now, evaluate tried to combine all available interfaces to satisfy
evaluation requests. For example, if Function implemented only
FunctionGradientHessian, FuncGradHess() calls were used also for
GradEvaluation and HessEvaluation requests. This was flexible but led to
a combinatorial explosion of cases. Due to a recent change Function must
now implement Gradient and/or Hessian interfaces when Methods needs
them, and the combined FunctionGradient and FunctionGradientHessian
interfaces became optional.  This together with ORable EvaluationTypes
enables considerable simplification of evaluate(). The combined
interfaces are still used whenever they are available and an appropriate
EvaluationType is requested.