Bisection: Change procedure to work toward the global minimum of the linesearch, not any local minimum.
The previous update rules could find a point of low function value, but stop the linesearch at a higher function value. This PR changes the behavior to have the linesearch stop at a point that has at least as low of a function value as the lowest returned point (see the discussion in issue 93). The point returned still meets the Wolfe conditions as originally specified. This should help stay in lower regions of attraction.