graph: fix graph copy functions
Previously, the copy functions would panic if an edge defined a node
prior to the node being added to the destination graph.

The fix here still allows the function to detect a node ID collision
between the source and destination graphs (the alternative approach to
not attempt to add nodes if present would not do this); this behaviour
is intentional.