Change Matrix definition to support transposes. Add Transpose type in support.
There is a need to support matrix transposes, for example the ability to perform A^T * B in addition to A * B. One way to support this is to add input booleans specifying if the matrix is transposed, as is done now in MulTrans. This does not scale well for many arguments and functions. This change adds a T() method to Matrix to support implicit transposing of matrices. This allows user to call, for example, Mul(A.T(), B) to perform A^T * B.

Please see the discussion in 123 for more information.

Changes to actual methods in order to efficiently support this new method type will come in future commits.

Fixes issue #123.

Updated comments

Fix documentation