Rename Linesearch and LinesearchMethod
Linesearch is a too nice name to be used for only a helper struct.
LinesearchHelper better transmits its purpose and leaves the name
Linesearch free for what used to be LinesearchMethod. LinesearchMethod
was a bit confusing because Method is the interface for any optimization
method and so the name suggested a relation that was not there.