mat64: drop TCopy method and associated interface
Depending on the desired behaviour, m.TCopy(a) can be replaced by m.Copy(a.T())
or m.Clone(a.T()). This significantly clarifies the intention over the
previous situation; TCopy's semantics were weird - it did not obey the
normal copy semantics (it allocated when the receiver was zero) and did
not obey normal clone semantics (it did not destroy the surrounding
matrix).