graph: split mutation API into smaller interfaces
graph.Mutable was previously used primarily to add nodes and edges, so
it is not sensible to require that edge and node deletion also be
implemented - there are add-only graph implementations that gain
possible optimisations by this change.

This change also potentially paves the way for multigraph addition with
an EdgeAdder interface.

Also drop the fictional distinction between the two Copy functions.