mat64: make both triangles when making random input
TriDense.Triangular returns true for upper on the zero value, so we were
not creating lower triangles. The change here is more strict than
absolutely necessary, but I think this is justified in testing.

We need to add the upper and lower cases to basicTriangular so that we
get both states for them as well.