graphs/gen: make duplication graph generation reproducible
The duplication graph algorithm depends on the order of nodes returned
by the Nodes and From methods, which are not guaranteed to be in a
consistent order. This impacts on the graphs generated. To allow it to
be used in benchmarking and other simulations, reproducibility must be
enforceable, so sort the nodes and depend on the random source for
randomness.