mat64: implement Binary(Un)Marshaler for Vector and Dense
This CL makes *mat64.Vector implement encoding.BinaryMarshaler and
encoding.BinaryUnmarshaler.
The (un)marshaling code for Dense and Vector is checking for
unreasonnable slice sizes (ie: it returns an error when the buffer
won't fit in the maximum slice size for the architecture.)
This CL also adds R/W tests for NaN, +Inf and -Inf for Dense and Vector.

Fixes gonum/matrix#337.
Fixes gonum/matrix#339.