Add Fortran-backed Airy function
This function adds the Airy function to mathext. This current implementation is backed using calls to Fortran for testing purposes. The tests confirm an exact numeric match between the Go code that is there with the Fortran calls to elementary functions (eps, zabs, etc.). It thus requires a fortran compiler and a post Go 1.6 implementation. The next commit will remove this dependence on Fortran at the cost of no longer an exact match due to small computation differences between the elementary function of both languages.