Remove Fortran dependence in Airy tests.
The first iteration of the code compared function output with the exact answer from Fortran. This is undesirable since it calls fortran at all, but beyond that the tests are also dependent on the compiler. This PR removes all calls to Fortran during actual execution and testing. To keep a test suite, make a copy of the code at this iteration exactly, so that moving forward we have a stable non-Fortran code to test changes against.