refactor hermiteInitialGuesses
The original implementation finds all of the initial guesses for every point. However, only parts of these initial guesses are used, so this wastes computation and allocates extra memory. This refactors the code to only compute the guesses that are actually used, which saves computation time and removes all memory allocation