Use Travis Caching to Speed up OpenBLAS
Travis has a caching feature, where you can "persist" a directory across
CI calls, detailed at https://docs.travis-ci.com/user/caching/

This change creates a cache in the OpenBLAS.cache/ directory (only
available on travis) which holds the OpenBLAS and netlib libraries and
headers used.  The installation script checks to see if the cache
contains the latest version of OpenBLAS by checking the commit id, and
if there is a newer commit on the OpenBLAS git repo, then the cache is
cleared and rebuilt.  There is no handling of the netlib cblas
versioning at the moment.

Each item on the built matrix gets a different cache, so we will still
build OpenBLAS at least as many times as we have go versions in the
build matrix, but after that they won't build again until OpenBLAS
updates.