native: add indicator of convergence to returned values from Dlahqr
Previously, the zero value of the returned integer info was used to indicate
success. This is wrong because it is a 0-based index unlike the reference LAPACK
where it is a 1-based index and therefore 0 is available for such purpose. This
commit adds `converged bool` to the returned values from Dlahqr.