native: check input slices to Dlaqr2, Dgehrd and Dormhr only after workspace query
so that it is possible to pass all slices as nil to those functions. A workspace
query should depend only on the input sizes, the matrices and vectors must
not be accessed anyway or may not be available at the call site. Moreover, often
the reference LAPACK uses arbitrary placeholders for arrays and relies on the
fact that no checking is done.