Initial CONTRIBUTING.md proposal (#26)
Initial CONTRIBUTING.md

Providing contribution guidelines helps newcomers be more productive, and helps firm up the project goals.  For #25.

The Gonum-dev group was also notified.  There are still some post-LGTM comments to address in a followup pull request.  I'm closing this so that we have something in place that we start to improve, instead of continuing to improve something that is already generally agreed upon without merge.