Optimize NormalQuantile
This commit makes the coefficient lists arrays instead of slices, changes the iteration order so the order of indexing is in the right direction, and manually unrolls the iteration loops. There are also a bunch of minor code cleanups. This reduces the number of lines of code, reduces the error surface, and cuts the evaluation time in half.  These net positives outweigh the reduced legibility in the rational approximations themselves.