graph/formats/gexf12: new package for handling GEXF1.2 graph description files
This was largely written by aqwari.net/xml/cmd/xsdgen working on the XSD
files available from the gephy/gexf specs repository[1], with some help;
the specs use includes which cause xsdgen to fail, so I manually
in-lined the includes.

The generated code was then edited to make optional attributes exist on
pointers, to reduce the length of labels to within reasonably normal Go
idiom and removing unnecessary namespacing.

The package is versioned since the format appears to arbitrarily be
updated.

[1]https://github.com/gephi/gexf/tree/81ba4e7ccdc25631f836fc5caa4ed64ba5300379/specs/1.2draft