graph/{multi,simple}: harmonise code order
Previously there was an attempt to maintain some kind of semantic ordering,
but given the needed complexity of the API this has become untenable. So
just order according to the Godoc — alphabetical for exported methods and
functions. For unexported functions, use judgement.

One method, Has on multi.DirectedGraph, that should not have been present is
deleted.