Listtypes (#714)
* mat: Add missing types to list test

With the recent additions of Diagonal and TriBanded, there were some matrix types that were not added to the list of testing types. There were additionally some older types that had been left out. This PR completes the set of types, their transposes, and their basic counterparts to the set of matrices. Changes are made to the makeNew etc. routines in order to support this.

This also changes the behavior of Triangle to return the embedded TriKind even if the matrix IsZero (before we intentionally corrupted the value). We document that the value cannot be trusted, so it shouldn't matter, but this way is simpler, and allows the TriKind to be used in list_test to specify matrices with different TriKind.

Fixes #266