lapack/testlapack: don't test unsupported matrix sizes in DlantrTest
The documentation for DLANTR in Reference-LAPACK says that the input
matrix should not be tall if upper triangular and not wide if lower
triangular. In most cases providing such "invalid" sizes is harmless and
DLANTR works correctly. However, when computing the inf-norm of lower
triangular matrices it currently causes an out-of-bound write if the
work array is shorter than the number of columns. Even if the reference
fixes this, we cannot assume when or if at all other LAPACK
implementation providers include it (both OpenBLAS and MKL have this
issue, obviously OpenBLAS being much easier to fix). Therefore, the
restriction on matrix sizes will have to stay in the reference
documentation and we should exlude them from our testing.