mat: calculate Q lazily when calling QR.ToQ
When a matrix is very tall, calculating Q will currently allocate a large Q at
the end of the factorisation, even if it is not going to be used. The eager
calculation was intended to prevent repeated re-calculation of Q when it is
used. So move the Q calculation to ToQ, but make it conditional on the stored
Q value being empty, and empty Q at the end of the factorisation.